backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "app"},
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.95.2"
uvicorn = "^0.23.2"
json-logging = "^1.3.0"
requests = "^2.27.1"
pytest = "^7.0.1"
pydantic-mongo = "^1.0.1"
langchain = "^0.1.16"
pymongo-inmemory = "^0.2.13"
anyio = "^3.7.0"
httpx = "^0.24.1"
python-multipart = "^0.0.6"
minio = "^7.1.15"
boto3 = "^1.28.2"
openai = "^1.17.0"
pypdf = "^3.15.5"
chromadb = "^0.4.9"
tiktoken = "^0.6.0"

newspaper3k = "^0.2.8"
langchainhub = "^0.1.13"
unstructured = "^0.10.16"
pybars3 = "^0.9.7"
twilio = "^8.9.1"
huggingface-hub = "^0.19.4"
griptape = "^0.24.2"
langchain-openai = "^0.1.3"
langchain-community = "^0.0.32"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
